Add new configuration to ModuleHeader step
Previously the formatter of module header would always add open
bracket of the export list on the next line after the module name.

With this change we give user ability to customize this behavior. The
open bracket ends up either on the same line as module name or on the
next line.

For backward compatibility we keep the 'next_line' a default behavior.