Add new configuration to ModuleHeader step (#357)
Add a new configuration to ModuleHeader step

Previously the formatter of module header would always add open
the bracket of the export list on the next line after the module name.

With this change, we give the user the ability to customize this behavior. The
open bracket ends up either on the same line as module name or on the
next line.

For backward compatibility, we keep the 'next_line' a default behavior.