Fix Nix build, bumping haskell.nix (#397) (#398)
* Fix Nix build, bumping haskell.nix (#397)

Fixes `nix-shell` and `nix-build` for the repo, which was failing with a dependency error (#397).

This PR takes two steps to fix this:

  1. It bumps the version of `haskell.nix` we use, switching to GHC 8.10.7 and Nixpkgs 21.11 so that we still use the haskell.nix cache

  2. After 1., I ran into an error where haskell.nix was using Cabal-3.2.1.0 even though `stylish-haskell.cabal` needs Cabal >= 3.4. To fix this, I had to explicitly mark the Cabal library as "reinstallable"—otherwise, haskell.nix uses the exact version that comes with GHC. (See [haskell-nix issue #1337][1] for details.)

[1]: https://github.com/input-output-hk/haskell.nix/issues/1337

I tested this change locally. Both `nix-build` and `nix-shell` succeeded without needing to build GHC from source.

* Set shell.nix to explicitly import haskell-language-server = 1.6.1.1

The version of haskell-language-server from pkgs ran into a bug without an explicit hie.yaml. Importing a specific version fixes that problem.