Rename constructors and related parser strings of LongListAlign
The documentation of these doesn't quite reflect the implementation
and it seems to cause some confusion.

This patch tries to address it by renaming `LongListAlign`
constructors to better reflect what they do and fix their
documentation in `stylish-haskell.yaml`.

There is no change in functionality but it introduces a breaking
change for users using the Haskell interface. It also adds
a (lousy) concept of config option deprecation in form of documenting
the deprecated variants in favor of new names.

Related to #462